+++ wfapi/features/__init__.py
@@ -1,4 +1,10 @@
 # -*- coding: utf-8 -*-
+from ..exception import WFUnsupportedFeature as _WFUnsupportedFeature
 
 from .deamon import *
 from .weak import *
+
+try:
+    from .autologin import *
+except _WFUnsupportedFeature:
+    pass